+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+## 1.0.0
+- Plugin updated to flutter 3.3.0
+- Added support for new Android plugins APIs
+
 ## 0.0.2
-Added auto-publish with Github Actions
+- Added auto-publish with Github Actions
 
 ## 0.0.1
-Initial Release
+- Initial Release